requirements.txt
pandas
kafka-python
elasticsearch
vnstock